requirements.txt
Flask==2.3.3
aiohttp==3.9.5
PyYAML==6.0.1
